Install and configure PHPCode_Sniffer
In order to automatically check code style we need a tool. The
PHPCode_Sniffer is a relatively wide used tool, so I've installed it.
Shopware recommends to use PSR-1 and PSR-2 standards.
@see https://developers.shopware.com/developers-guide/coding-standards/#php-coding-standard
> For PHP Code all contributions should use the PSR-1: Basic Coding 
> Standard and PSR-2: Coding Style Guide.
These will be check my PHPCode_Sniffer too.
Currently only the ./src folder is checked. In the future the 
./tests folder might be needed to be checked too.