Intercept "click" Events instead of "submit" events.
Current solution seems to have some problems with the way shopware handles the submit.
There is a custom submit listener for every wform and so if the form is being sent, our plugin
enters a race condition with shopware's submit listener. Sometimes the one from shopware has 
higher priority and the form essentially gets sent.
Which results in wrong addresses landing in the database.
So in order to solve this problem I made a change on when we intercept a submit. Instead
of actuall "submit"-event we intercept a "click"-event before that. This seems to give us more 
stable and predictable results.