Improved compatibility with 6.5
in this commit I addresses some of the issues with 6.5 shopware.
First of all I rewrote the routes definition into php. This way,
as some kind of workeround, we can avoid using php attributes and
therefore keep the backward compatibility with 7 php version.
Other issues involved the problem with missing jquery or $
Looks like shopware has a cut down version of it with some missing
methods. So I added a check for jquery, val and trigger methods.
If they miss I just go back to direct value assignment.
This helped to solve the first appearing issues.