Create salutation mapping list on server side
In order to be able to correctly select the right option
in the salutation dropdown there is a need to create a mapping
between values that endereco api can deliver and values
that can be used in shopware.
In this commit I added a simple iteration, that goes through
a list of salutations from the shopware database and creates
mapping for those, that can be supported by endereco api.
The mapping is a mapping between internal shopware id and
the salutation code from endereco api.
Additionaly in this commit I provide those mapping to the frontend and
modifiy resolvers to use them when writing/reading to or from
the salutation select element.