Extend the waiting for free space logic
In Shopware 6 there are modals that the system inself can generate.
This collides with modals from endereco, as endereco js-sdk has to figure out
whether to display a modal or not.
On top of that the modals have semi levels.
Which means that on the first level, you might have two modals competing with each other,
but then in the process of editing an address from the first modals you
would open a shopware modal which in turn can trigger another modal (if you
would make a speccil error while editing the address), which should be displayed
before the original second modal.
It quire confusing, i know.
In this fix we add two things:
1. The intent why a form is loaded. If the intent is "to edit", then no modal
 be displayed initialy.
2. For some EAO object we change the waitForFreeSpace function in order to also
 cnsider their level