Fix the bug with not nullable extension fields
With the 1.4.0 we made housenumber and street not nullable. However the Doctrine, that Shopware ist using internally, seems to transform an empty string into a null. So when the user tries to save an address
that doesn't have a house number, therefore the "housenumber" is "", it gets transformed into a "NULL", and causes an SQL Exception, when tryging to write a NULL to a non nullable field.

So to fix it, we made the fields nullable.