Remove changes from migration and prevent settings loading
One of the migrations turned out to be problematic, so in order to prevent the damage we
removed the sql command in that migration. The following migration repairs everything, so there should be
no more issues with it.

We also added a blocked, that would prevents loading of country/regions lists for pages, that don't need them.