Read submitted data von address metadata from requestbag if avaiable
If a guest submits an address, but there is a serverside validation error, for example
the email is already in use, then the user ist redirected to the original page. In the current
implementation the metadata of the address would be lost, because they are not saved in the
extension yet. Which in turn can lead to a situation, when an already checked address doesnt
produce a correction modal in the next step in the checkout.

So in order so ensure, that metadata is not lost, we first check im request bag contains data,
and only then in the extension.

Ref: S6A-209