Delete previos container
IF playground starts a container with particular name, it can not be created the next time, because
the name is taken. So in order to be able to execute playground.sh every time, the old container is#
deleted (if it exists) before we attempt to start it.

I also added the 6.6.2.0 version of shopware in the list, however there is no dockeware image for it
at the moment. Gotta wait a bit, the code will work later.