Set default value to extension fields if no value is provided in payload
IN checkout its possible to select and apply a correction of an address to the billing or
the shipping address. However the controller expected that the fields "enderecoStreet" and "enderecoHousenumer"
are set, which is not the case and would trigger 500 error under some circumstances.

In this commit I would provide empty string to the extestension fields if the fields in
ajax payload are not existing, which solves the problem. After reloading the page, the fields for
streetname and building number would be populated automatically.

Ref: S6A-215