Update js-sdk and modify country state reader filter
This version of js-sdk provides a reference of subscriber to readFilterCB (subdivisionCodeRead)
which is normally responsible for mapping of the local value to a value that is needed
by EAO. So this is a hack essentially. Before the value is mappen we check if the source of the value
is actually ready to provide value (if the select has more than 1 option). If its not ready (1 option)
then we read the value from a specific data attribute and use this value instead of the original one.
Same for country code.
This way the address that is rendered via ajax will not be marked as changed.

Ref: S6A-211