Fix sed syntax issue
sed expects different -i attribute depending on the os. Its different with macos comapred to linux. So in order to make sure it works
in both, we need to change th -i param to something that works in both. It creates some .bak files though, so those needs some cleaning.