Add optional processing for imported customers and customer addresses
If the customer list is imported through admin setting import/export, that triggers the logic from
endereco plugin, at least triggering the splitStreet and in some cases also the existing customer check.

This behavior is not consistent and causes address checks that some merchants might not want to pay, so it
should be optional and configurable. In case the merchant activates teh feature, all addresses from import list
will be validated, the streets will be splitted.

Red: S6A-227