Use null coalescing operator for $_SERVER
For request that are triggered from the backend, we try to set the referrer from $_SERVER['HTTP_REFERER']
which sometimes doesnt exist (the key itself doesnt exist), which would trigger a php warning.

Now in SHopware 6 php warning are handler by symfonys error handler, that starting from sw 6.6.4.0 and symfony 7.1
sees every php warning as error and logs it to error log. Which creates a lot of support tickets for us and
reduces the perceived quality of the product.

Ref: EST-433, EST-429