Copy endereco address extension data during cart conversion
The conversion of the cart into an order is the earliest event to copy validation data from customer address into order address. In fact the order address is not even created yet, so we just prepare the data including the data in the extension.
Once the order addresses are create, the extension with validation data will be created too.