Transform address extension definition into base class and create two implementations
Now that we want to save address statuses and other meta information both in the customer address and in the order address
it makes sense to have a base class for both of them to contain properties that appear in both cases. This should make future changes
a bit more secure.
To make the difference between the meta data for customer address and order address clear, the name of the customer address should be changed.
Both extension table names have a suffix "_gh", which is replaced with "_sw" in the store version, to prevent future collision between the github and store version.