Fix country and state selection criteria to select only active
A customer reported a problem with our plugin. She couldn't select Poland from a country list 
when Endereco plugin was active. After researching the problem, we found out, that Poland was created
two time in customer shop, the first one was deactivated, the second was active.
Endereco plugin used the deactivated country for country mapping table, the frontend used active country record for display, so it would not be possible to select a correct Poland in the options list via ID. the ID was from the deactivated country which was not in the list.
The solution is to filter out inactive countries and - while we are at it - also the states (possibly similar problem).

Ref: EST-493