Add custom fields to order to contain address validation data
To simplify access to the address validation data it might be useful to mirror the validation data into a custom field in order. It is saved and accessed differently
and might be easier to read from in some cases.
The customer field is always secondary and is updated when an order address entity extension is updated.
There should be different custom fields for billing and shipping addresses, because in case there is some kind of automation, there is a difference between a
wrong billing address and wrong shipping address.