Recalculate hash payload when the address extension is written
In order to invalidate the results of an address check later in case of the address being changed by third party there should be
a kind of signature that allows for matching between the address and the meta data in extension. Generate a payload that would be sent
to Endereco API as unique representation of the address from the perspective of metadata. For the same payload the metadata are supposed to be
the same too. If the original data would be changed, then the newlycalculated payload would be different to the saved payload, hence the metadata in extension
can be discarded.