Apply address results also when address is correct
Because the API can accept streetFull but returns separated street and house number, its possible, that
the information about order is lost, which allows for a mistake in the address to stay unnoticed.
So as a quickfix we just rebuild the street for the prediction, that by definition contains an address, that is 100% correct.
The rebuild logic is considering different position of house number and street for some countries.
The problem might appear in the missing processing of the additional info, which might be removed from default "street" field, but not added to "additionalInfo".
Because now we work with address signatures, the automatic aplication might spoild the signature, making it invalid. so before we are done with the insurance logic, we check if signature is still valid. if not, the address is checked again. This way we ensure, that the status is really set.
Because after the check we optionally apply the results to customer address, this might trigger a session accounting (entity write subscriber), which is faulty. So for the time we are processing the insurances the accounting is deactivated (the sessions are coutned though), and is activated again when all the insurances are done. This is when the final doAccounting happens too.

Ref: S6A-241