[WIP] Add support for additionalInfo
In order to process an address properly, we need to consider the additional info field too. Sometimes the "street" in shopware 6 shop contains information, that doesnt belong to street or couldn't be printed on the label because of the field size, so it has to be in the line over the street. Usually the additionalAddressLine(1|2) is used for that.

Because the API returns (eventually) the additional info in one field, but shopware 6 has two possible fields to save it, a mapping service takes over deciding which field to use. It relates mostly to customer addresses in the storefront and hasn't been tested for headless or Admin API access.

The streetIsSplit Insurance for orderAddresses has been removed, because we dont know if we can change the order address safely (street split can rewrite the "street" and one of the additionalAddressLine fields).

Scope of Changes:
- Added new mapping service for additional address info
- Modified address processing logic in controllers and subscribers that proccess customer address
- Removed streetIsSplit insurance from order addresses (for now)

Known Limitations:
- Not tested with headless or Admin API access
- Customer address modifications might need additional safety checks

Test Plan:
Manual tests:
- Test creating and saving addresses in Registration, Checkout and Addresses (for logged in users)
- Use different storefront setting (no additionaAddressLine, additionaAddressLine1 active, additionaAddressLine2 active, both active)
- Observe the payload send to API. Depending on the storefront settings it should or should not contain the "additionalInfo"

Ref: S6A-246