Move splitStreet into a service and add cache through a decorator
As street splitting is a process done by Endereco, caching is useful to avoid repeating requests. The `splitStreet` method was moved from the `EnderecoService` to the `StreetSplitter` service. This service was then decorated with a cached version that uses a Symfony cache defined by this plugin. The `getAgentInfo` method, the `preparePayload` method, the `getPluginVersion` method and the `generateRequestHeaders` method were moved from the `EnderecoService` to separate services in order to avoid circular dependencies between the `StreetSplitter` and the `EnderecoService`.

Co-authored-by: Ilja Weber <ilja@endereco.de>