Set is frontend flag to additional context service
We found out that some operations that should normally run only in the storefront, are actually executed in backend. Possibly in pickware ERP too.
This is a big problem, because it can potentialyl create a high usage of our API that the customer would not be willing to pay for. to solve this problem
a service was created that contains a flag that is set to true when storefront page is loading. This criteria is supposed to be typical only
for the actual storefront and allow us to limit those critical operations to only storefront.

Ref: EST-429