Add a datastucture to represent and validate update payload
In DEV-60 we documented a bug, in which in case of a very long string,
that our web API returns, we break the shop. Essentially the plugin tries
to write a string to database fields, that cannot accept the string of that length.

Currently it looks like this:
  Faulty Data --> Repository

Despite the obvios fix for the API, it's clear that the client cannot trust
the API completely, as the API is not aware of client system limitations.

In order to address this problem, a new data structure was created. It's composed of
customer address update payload and customer address extension update payload.
Those payloads represent the data that will go to database at some point and ensure
that the limits of the client system are respected.

So it looks like this:
  Faulty Data --> Payload Model (makes Correct Data) --> Repository

This fixes the breaking of the shop due too some strings being too long for update.
This does not fix or address any performance or endless loop issues.

Ref: https://mobilemojo.atlassian.net/browse/DEV-60?focusedCommentId=38153