Ensuring MySQL 8.4 Compatibility
Removed the constraint in CREATE if the MySQL-Version is higher or
equal than 8.4. We only adapt the non-executable migrations for
known failures in all database systems currently officially
supported by Shopware (MySQL > 8.0.22 and MariaDB > 10.11.6
or MariaDB> 11.0.4, which can be found here:
https://developer.shopware.com/docs/guides/installation/requirements.html#sql)
Migration1731623484 is altered only for MySQL-Versions >= 8.4
Since Migration1731623484 may already have been executed in systems
in the field, the same migration becomes executable for the remaining
systems that previously threw an error.
This should be the preferred implementation in order to guarantee
identical states on customer systems after identical migrations.
Subsequently we standardize all databases back to the same data model
in Migration1753461221 which is tested to run in all officially
supported configurations.

Other database technologies were not considered due to the scope of
the current branch.

Ref:#88